Added support for multiple servers and paths
