+++ agent/ServiceConfig.cs
@@ -7,11 +7,16 @@ public sealed class AuthenticationConfig
     public required string Password { get; set; }
 }
 
-public sealed class ServiceConfig
+public sealed class ServerConfig
 {
     public required AuthenticationConfig Authentication { get; set; }
 
     public required Uri ServerUrl { get; set; }
 
-    public required string WatcherPath { get; set; }
+    public required List<string> WatcherPaths { get; set; }
+}
+
+public sealed class ServiceConfig
+{
+    public required List<ServerConfig> Servers { get; set; } = [];
 }
