+++ agent/appsettings.Development.json
@@ -6,11 +6,17 @@
     }
   },
   "ServiceConfig": {
-    "ServerUrl": "http://192.168.2.12:5000/",
-    "WatcherPath": "D:\\temp\\pdb_uploads",
-    "Authentication": {
-      "Username": "admin",
-      "Password": "admin"
-    }
+    "Servers": [
+      {
+        "ServerUrl": "http://192.168.2.12:5000/",
+        "Authentication": {
+          "Username": "admin",
+          "Password": "admin"
+        },
+        "WatcherPaths": [
+          "D:\\temp\\pdb_uploads"
+        ]
+      } 
+    ]    
   }
 }
